requirements.txt
pypdf2
tabulate
tqdm
pandas
google-generativeai